refactor: Update migrations to use integer columns for monetary values
Changes balance, initial_balance, and amount columns to bigInteger type
for storing monetary values as cents. Removes _cents suffix from column
names as the MoneyCast handles the conversion transparently.

+++ database/migrations/2025_08_09_161841_create_accounts_table.php
@@ -15,8 +15,8 @@ return new class extends Migration
             $table->enum('type', ['bank', 'cash', 'credit_card', 'loan']);
             $table->string('icon')->nullable();
             $table->string('color')->default('#3b82f6'); // Blue
-            $table->decimal('balance', 19, 4)->default(0);
-            $table->decimal('initial_balance', 19, 4)->default(0);
+            $table->bigInteger('balance')->default(0);
+            $table->bigInteger('initial_balance')->default(0);
             $table->string('currency', 3)->default('USD');
             $table->string('account_number')->nullable();
             $table->text('notes')->nullable();
